refactor(utils): split logging out and export utils
