+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './GiftedChat'
 export * from './Constant'
+export * from './utils'
