+++ src/utils.ts
@@ -39,13 +39,3 @@ export function isSameUser(
     diffMessage.user._id === currentMessage.user._id
   )
 }
-
-const styleString = (color: string) => `color: ${color}; font-weight: bold`
-
-const headerLog = '%c[react-native-gifted-chat]'
-
-export const warning = (...args: any) =>
-  console.log(headerLog, styleString('orange'), ...args)
-
-export const error = (...args: any) =>
-  console.log(headerLog, styleString('red'), ...args)
